+++ lib/src/column_sort.dart
@@ -1,9 +1,9 @@
-import 'package:davi/src/sort_order.dart';
+import 'package:davi/src/sort_direction.dart';
 
 /// Describes the column sort.
 class ColumnSort {
-  ColumnSort({required this.columnIndex, required this.order});
+  ColumnSort({required this.columnIndex, required this.direction});
 
   final int columnIndex;
-  final TableSortOrder order;
+  final DaviSortDirection direction;
 }